🚚 Rename Routes from js to tsx
